Add files via upload
Initial Very Rough Commit of code used to do certain sample regressions documented in Table 1 of Chatfield et al., Atmospheric Chemistry and Physics, 2020.  Most likely this code is only a general guide to what a user must do. A guide to the processing steps isn in PM2.5_Procedures_General_Notes.txt  NOTE:  Much of the dataset generation is idiosyncratic to California, a special aircraft investigation in 2013, and the 2019 version of the MAIAC aerosol algorithm, Version 6: See Lyapustin references in the A C P publication, currently in finalized draft form.